requirements.txt
certifi
emoji
PySocks
python-telegram-bot
clarifai
dnspython
pymongo
